# backend/build/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Gender": {
				"enum": [
					"m",
					"f",
					"o"
				],
				"type": "string"
			},
			"Role": {
				"enum": [
					"super_admin",
					"artist_manager",
					"artist"
				],
				"type": "string"
			},
			"Pick_IUser.Exclude_keyofIUser.id-or-created_at-or-updated_at__": {
				"properties": {
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"dob": {
						"type": "string"
					},
					"gender": {
						"$ref": "#/components/schemas/Gender"
					},
					"address": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					}
				},
				"required": [
					"first_name",
					"last_name",
					"email",
					"password",
					"phone",
					"dob",
					"gender",
					"address",
					"role"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_IUser.id-or-created_at-or-updated_at_": {
				"$ref": "#/components/schemas/Pick_IUser.Exclude_keyofIUser.id-or-created_at-or-updated_at__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"IUserInput": {
				"$ref": "#/components/schemas/Omit_IUser.id-or-created_at-or-updated_at_"
			},
			"Partial_IUserInput_": {
				"properties": {
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"dob": {
						"type": "string"
					},
					"gender": {
						"$ref": "#/components/schemas/Gender"
					},
					"address": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"IUserOptional": {
				"$ref": "#/components/schemas/Partial_IUserInput_"
			},
			"IUser": {
				"properties": {
					"id": {
						"type": "string"
					},
					"first_name": {
						"type": "string"
					},
					"last_name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"dob": {
						"type": "string"
					},
					"gender": {
						"$ref": "#/components/schemas/Gender"
					},
					"address": {
						"type": "string"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"created_at": {
						"type": "string"
					},
					"updated_at": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"first_name",
					"last_name",
					"email",
					"password",
					"phone",
					"dob",
					"gender",
					"address",
					"role",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "Enter token for User"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/user/register": {
			"post": {
				"operationId": "RegisterUser",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserInput"
							}
						}
					}
				}
			}
		},
		"/user/login": {
			"post": {
				"operationId": "LoginUser",
				"responses": {
					"200": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"password": {
										"type": "string"
									},
									"email": {
										"type": "string"
									}
								},
								"required": [
									"password",
									"email"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/user/update/{userId}": {
			"patch": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [
					{
						"jwt": [
							"super_admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserOptional"
							}
						}
					}
				}
			}
		},
		"/user/delete/{userId}": {
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [
					{
						"jwt": [
							"super_admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/user": {
			"get": {
				"operationId": "GetUser",
				"responses": {
					"200": {
						"description": "Fetched",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [
					{
						"jwt": [
							"super_admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "pageNumber",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "rowsPerPage",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/user/userDetails": {
			"get": {
				"operationId": "GetUserDetails",
				"responses": {
					"200": {
						"description": "Fetched",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/IUser"
										},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"user"
				],
				"security": [
					{
						"jwt": [
							"super_admin",
							"artist",
							"artist_manager"
						]
					}
				],
				"parameters": []
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:8000/"
		}
	]
}